+++ CHANGELOG.md
@@ -1,11 +1,12 @@
 # Change Log (vs-deploy)
 
-## 9.11.0 (May 29th, 2017; settings)
+## 9.11.0 (May 29th, 2017; settings and placeholders)
 
 * added `alwaysShowPackageList` setting that indicates if package list is shown, even if there is only 1 entry
 * added `alwaysShowTargetList` setting that indicates if target list is shown, even if there is only 1 entry
 * added `username` and `password` settings for [HTTP target operations](https://github.com/mkloubert/vs-deploy/wiki/targetoperations#http-)
 * added build-in [placeholders](https://github.com/mkloubert/vs-deploy/wiki/values) `EOL`, `hostName` and `tempDir`
+* added `startBuildTask` that can defines if `workbench.action.tasks.build` (build task) should be run on startup or not
 
 ## 9.10.0 (May 28th, 2017; new list target)